refactor(filters): move sync-time filtering out of commands
