+++ src-tauri/src/filters/mod.rs
@@ -1,2 +1,3 @@
 pub mod engine;
 pub mod rules;
+pub(crate) mod service;
